Backward and Forward Chaining decision trees
Decision tree is a diagram which helps to visualize all the factors that must be considered for taking a decision. Decision trees have decision nodes with questions which branch out to further nodes, and finally reach a conclusion node. Working of a human mind while taking a decision is mimicked here. In our system, we have used a backward chaining and forward chaining tree to mimic this decision-making process.